bysc.h :: fix BYSC_LINE_LIMIT array specification
the array size specification that is bound to 'BYSC_LINE_LIMIT' in bysc.h, was [ last ] defined as '(int)(((unsigned int)(~0) / 2))', redefined as '(int)(((unsigned int)(~0) >> 2))'.

REASON :: defining an array size as 0x7FFFFFFF ( larger than 2GiB ) causes a break in the link between the program and the standard C libraries on the latest versions on macOS. size reduced to 0x3FFFFFFF (1GiB). good fix for macOS, it now compiles and runs fine on macOS 12.1, now i must reconvene with design specifications i wrote for command parser in 'command.c'